Move tx and reject to own files
